Update config structure

+++ pymovis/learning/dataset.py
@@ -1,93 +1,36 @@
 import os
 import pickle
-import torch
-from torch.utils.data import Dataset
-import numpy as np
 import random
 
-from pymovis.utils import util
+import torch
+from torch.utils.data import Dataset
 
-class WindowDataset(Dataset):
-    def __init__(self, path):
-        self.path = path
-        self.windows = []
-        for f in sorted(os.listdir(path)):
-            if f.endswith(".txt"):
-                self.windows.append(os.path.join(path, f))
-        
-        with open(os.path.join(path, "skeleton.pkl"), "rb") as f:
-            self.skeleton = pickle.load(f)
-    
-    def __len__(self):
-        return len(self.windows)
-    
-    def __getitem__(self, idx):
-        res = np.loadtxt(self.windows[idx], dtype=np.float32)
-        return torch.from_numpy(res)
-    
-    def get_mean_std(self, dim):
-        # load and return mean and std if they exist
-        mean_path = os.path.join(self.path, "mean.pt")
-        std_path = os.path.join(self.path, "std.pt")
-        if os.path.exists(mean_path) and os.path.exists(std_path):
-            mean = torch.load(mean_path)
-            std = torch.load(std_path)
-            return mean, std
 
-        # load all windows in parallel
-        indices = list(range(len(self)))
-        items = util.run_parallel_sync(self.__getitem__, indices)
+class MotionDataset(Dataset):
+    def __init__(self, train, config):
+        self.train  = train
+        self.config = config
 
-        # calculate mean and std
-        items = torch.stack(items, dim=0)
-        mean = torch.mean(items, dim=dim)
-        std = torch.std(items, dim=dim) + 1e-6
-
-        # save mean and std
-        torch.save(mean, mean_path)
-        torch.save(std, std_path)
-
-        return mean, std
-
-    def get_feature_dim(self):
-        return self[0].shape[-1]
-
-""" Dataset pair of (X, Y) where X is a set of motion features and Y is a set of environment features """
-class PairDataset(Dataset):
-    def __init__(self, base_dir, train, X_name="motion", Y_name="envmap", window_size=50, window_offset=20, fps=30, sparsity=15, size=200, top_k_samples=10):
-        self.base_dir = base_dir
-        self.train    = train
-        self.X_name   = X_name
-        self.Y_name   = Y_name
-
-        self.window_size = window_size
-        self.window_offset = window_offset
-        self.fps = fps
-        self.sparsity = sparsity
-        self.size = size
-        self.top_k_samples = top_k_samples
-
-        with open(os.path.join(base_dir, X_name, f"{'train' if train else 'test'}_length{window_size}_offset{window_offset}_fps{fps}.pkl"), "rb") as f:
-            self.X = pickle.load(f)["features"]
-        with open(os.path.join(base_dir, Y_name, f"{'train' if train else 'test'}_length{window_size}_offset{window_offset}_fps{fps}_sparsity{sparsity}_mapsize{size}_top{top_k_samples}.pkl"), "rb") as f:
-            self.Y = pickle.load(f)
-        with open(os.path.join(base_dir, X_name, "skeleton.pkl"), "rb") as f:
+        with open(os.path.join(self.config.dataset_dir, f"{'train' if train else 'test'}_{self.config.motion_pklname}"), "rb") as f:
+            self.data = pickle.load(f)["features"]
+        with open(os.path.join(self.config.dataset_dir, "skeleton.pkl"), "rb") as f:
             self.skeleton = pickle.load(f)
         
-        if self.X.shape[0] != self.Y.shape[0]:
-            raise ValueError("X and Y must have the same number of samples")
-        
+        print("Motion dataset: ", self.data.shape)
+    
     def __len__(self):
-        return len(self.X)
+        return len(self.data)
     
     def __getitem__(self, idx):
-        rand = random.randint(0, self.top_k_samples - 1)
-        return torch.from_numpy(self.X[idx]).float(), torch.from_numpy(self.Y[idx, rand]).float(), rand
+        return torch.from_numpy(self.data[idx]).float()
     
-    def get_motion_statistics(self, dim):
+    def shape(self):
+        return self.data.shape
+
+    def statistics(self, dim):
         # load and return mean and std if they exist
-        mean_path = os.path.join(self.base_dir, self.X_name, "mean.pt")
-        std_path  = os.path.join(self.base_dir, self.X_name, "std.pt")
+        mean_path = os.path.join(self.config.dataset_dir, f"motion_mean_length{self.config.window_length}_offset{self.config.window_offset}_fps{self.config.fps}.pt")
+        std_path  = os.path.join(self.config.dataset_dir, f"motion_std_length{self.config.window_length}_offset{self.config.window_offset}_fps{self.config.fps}.pt")
 
         if os.path.exists(mean_path) and os.path.exists(std_path):
             mean = torch.load(mean_path)
@@ -97,23 +40,93 @@ class PairDataset(Dataset):
         if not self.train:
             raise ValueError("Mean and std must be calculated and saved on training set first")
 
-        # load all windows in parallel
-        indices = list(range(len(self)))
-        X       = [self.__getitem__(i)[0] for i in indices]
-        X       = torch.stack(X, dim=0)
-        
-        # calculate mean and std
+        # load motion features and calculate mean and std
+        X    = torch.stack([self[i] for i in range(len(self))], dim=0)
         mean = torch.mean(X, dim=dim)
         std  = torch.std(X, dim=dim) + 1e-8
 
         # save mean and std
         torch.save(mean, mean_path)
-        torch.save(std,  std_path)
+        torch.save(std, std_path)
 
         return mean, std
+
+class EnvmapDataset(Dataset):
+    def __init__(self, train, config):
+        self.train  = train
+        self.config = config
+
+        with open(os.path.join(self.config.dataset_dir, f"{'train' if train else 'test'}_{self.config.envmap_pklname}"), "rb") as f:
+            self.data = pickle.load(f)
+        
+        # reshape to (B * top_K, T, D)
+        self.data = self.data.reshape(-1, self.data.shape[2], self.data.shape[3])
+        print("Envmap dataset: ", self.data.shape)
     
+    def __len__(self):
+        return len(self.data)
+    
+    def __getitem__(self, idx):
+        return torch.from_numpy(self.data[idx]).float()
+
+    def shape(self):
+        return self.data.shape
+
+    def statistics(self, dim):
+        # load and return mean and std if they exist
+        mean_path = os.path.join(self.config.dataset_dir, f"envmap_mean_{os.path.splitext(self.config.envmap_pklname)[0]}.pt")
+        std_path  = os.path.join(self.config.dataset_dir, f"envmap_std_{os.path.splitext(self.config.envmap_pklname)[0]}.pt")
+
+        if os.path.exists(mean_path) and os.path.exists(std_path):
+            mean = torch.load(mean_path)
+            std  = torch.load(std_path)
+            return mean, std
+        
+        if not self.train:
+            raise ValueError("Mean and std must be calculated and saved on training set first")
+
+        # load envmap features and calculate mean and std
+        X    = torch.stack([self[i] for i in range(len(self))], dim=0)
+        mean = torch.mean(X, dim=dim)
+        std  = torch.std(X, dim=dim) + 1e-8
+
+        # save mean and std
+        torch.save(mean, mean_path)
+        torch.save(std, std_path)
+
+        return mean, std
+
+""" Paired dataset for motion and envmap features """
+class PairDataset(Dataset):
+    def __init__(self, train, config):
+        self.train  = train
+        self.config = config
+
+        self.motion_dset   = MotionDataset(train, config)
+        self.envmap_dset   = EnvmapDataset(train, config)
+
+        if self.motion_dset.shape()[0] * self.config.top_k != self.envmap_dset.shape()[0]:
+            raise ValueError(f"Motion and envmap datasets must have the same data samples, but got {self.motion_dset.shape()[0]} and {self.envmap_dset.shape()[0]}")
+        if self.motion_dset.shape()[1] != self.envmap_dset.shape()[1]:
+            raise ValueError(f"Motion and envmap datasets must have the same length, but got {self.motion_dset.shape()[1:]} and {self.envmap_dset.shape()[1:]}")
+        
+    def __len__(self):
+        return len(self.motion_dset)
+    
+    def __getitem__(self, idx):
+        rand = random.randint(0, self.config.top_k - 1)
+        x = self.motion_dset[idx]
+        y = self.envmap_dset[idx * self.config.top_k + rand]
+        return x, y, rand
+    
+    def motion_statistics(self, dim):
+        return self.motion_dset.statistics(dim)
+    
+    def envmap_statistics(self, dim):
+        return self.envmap_dset.statistics(dim)
+
     def motion_dim(self):
-        return self.X.shape[-1]
+        return self.motion_dset.shape()[-1]
     
     def env_dim(self):
-        return self.Y.shape[-1]
+        return self.envmap_dset.shape()[-1]
